+++ src/errors.rs
@@ -22,9 +22,13 @@ macro_rules! impl_error {
 
 #[cfg_attr(feature = "std", derive(Error))]
 #[derive(Clone, Debug, PartialEq)]
+#[non_exhaustive]
 pub enum ArithmeticError {
     Overflow,
     DivisionByZero,
+    /// When someone tries to use operand out of the set of departure of the function.
+    /// E.g.: when you try to compute the square root of a negative number.
+    DomainViolation,
 }
 
 impl ArithmeticError {
@@ -32,6 +36,7 @@ impl ArithmeticError {
         match self {
             Self::Overflow => "overflow",
             Self::DivisionByZero => "division by zero",
+            Self::DomainViolation => "domain violation",
         }
     }
 }