+++ src/i256/mod.rs
@@ -2,7 +2,9 @@ use core::cmp::{Ordering, PartialOrd};
 use core::convert::TryFrom;
 use core::ops::{Div, Mul, Neg, Sub};
 
-use crate::ArithmeticError;
+use crate::ops::sqrt::Sqrt;
+use crate::ops::{One, RoundMode, RoundingSqrt, Zero};
+use crate::{ArithmeticError, ConvertError};
 
 const TOTAL_BITS_COUNT: usize = 256;
 const UINT_CHUNK_BITS_COUNT: usize = 64;
@@ -22,6 +24,7 @@ pub struct I256 {
 impl I256 {
     pub const I128_MAX: Self = Self::from_i128(i128::MAX);
     pub const I128_MIN: Self = Self::from_i128(i128::MIN);
+    pub const U128_MAX: Self = Self::new(U256([u64::MAX, u64::MAX, 0, 0]));
     pub const MAX: Self = Self::new(U256([u64::MAX, u64::MAX, u64::MAX, !SIGN_MASK]));
     pub const MIN: Self = Self::new(U256([0, 0, 0, SIGN_MASK]));
 
@@ -47,6 +50,7 @@ impl I256 {
 impl Mul for I256 {
     type Output = Self;
 
+    #[inline]
     fn mul(self, rhs: Self) -> Self::Output {
         let lhs_was_negative = self.is_negative();
         let rhs_was_negative = rhs.is_negative();
@@ -67,6 +71,7 @@ impl Mul for I256 {
 impl Div for I256 {
     type Output = Self;
 
+    #[inline]
     fn div(self, rhs: Self) -> Self::Output {
         let lhs_was_negative = self.is_negative();
         let rhs_was_negative = rhs.is_negative();
@@ -86,6 +91,7 @@ impl Div for I256 {
 impl Sub for I256 {
     type Output = Self;
 
+    #[inline]
     fn sub(self, rhs: Self) -> Self::Output {
         let (x, _) = self.inner.overflowing_sub(rhs.inner);
         Self::new(x)
@@ -95,17 +101,18 @@ impl Sub for I256 {
 impl Neg for I256 {
     type Output = Self;
 
+    #[inline]
     fn neg(self) -> Self::Output {
         // Neg isn't defined for `I256::MIN` because on two's complement we always have one extra negative value.
         debug_assert_ne!(self, Self::MIN);
-        const U1: U256 = I256::from_i128(1).inner;
         // Overflow takes place when we negate zero.
-        let (x, _) = (!self.inner).overflowing_add(U1);
+        let (x, _) = (!self.inner).overflowing_add(Self::ONE.inner);
         Self::new(x)
     }
 }
 
 impl Ord for I256 {
+    #[inline]
     fn cmp(&self, other: &Self) -> Ordering {
         match (self.is_negative(), other.is_negative()) {
             (true, false) => Ordering::Less,
@@ -116,6 +123,7 @@ impl Ord for I256 {
 }
 
 impl PartialOrd for I256 {
+    #[inline]
     fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
         Some(self.cmp(other))
     }
@@ -135,6 +143,59 @@ impl TryFrom<I256> for i128 {
             return Err(ArithmeticError::Overflow);
         }
         Ok(i128::from(x.chunks()[0]) | (i128::from(x.chunks()[1]) << 64))
+    }
+}
+
+impl From<u128> for I256 {
+    fn from(x: u128) -> Self {
+        Self::new(x.into())
+    }
+}
+
+impl TryFrom<I256> for u128 {
+    type Error = ConvertError;
+
+    fn try_from(x: I256) -> Result<Self, Self::Error> {
+        if x > I256::U128_MAX || x < I256::ZERO {
+            return Err(ConvertError::Overflow);
+        }
+        Ok(u128::from(x.chunks()[0]) | (u128::from(x.chunks()[1]) << 64))
+    }
+}
+
+impl One for I256 {
+    const ONE: Self = Self::from_i128(1);
+}
+
+impl Zero for I256 {
+    const ZERO: Self = Self::from_i128(0);
+}
+
+impl RoundingSqrt for I256 {
+    type Error = ArithmeticError;
+
+    /// Integer square root of a non-negative integer S is a non-negative integer Q such that:
+    /// Floor: `Q ≤ sqrt(S)`
+    /// Ceil: `Q ≥ sqrt(S)`
+    #[inline]
+    fn rsqrt(self, mode: RoundMode) -> Result<Self, Self::Error> {
+        if self.is_negative() {
+            return Err(ArithmeticError::DomainViolation);
+        }
+        let lo = self.inner.sqrt()?;
+        let inner = match mode {
+            RoundMode::Floor => lo,
+            RoundMode::Ceil => {
+                if lo * lo == self.inner {
+                    lo
+                } else {
+                    // `sqrt` will always be closer to zero than `self` so overflow will never happen
+                    let (hi, _) = lo.overflowing_add(Self::ONE.inner);
+                    hi
+                }
+            }
+        };
+        Ok(Self::new(inner))
     }
 }